feat: add workload numbers on sidebar, remove show more option on individual workload filters and move the useWorkloadCounter hook to hooks file
